// DXC_Frontend/src/api/models/iconCategoryDto.ts
/**
 * Generated by orval v7.11.2 🍺
 * Do not edit manually.
 * DXC_Core.API
 * OpenAPI spec version: 1.0
 */

export interface IconCategoryDto {
  publicId?: string
  /**
   * @minLength 1
   * @maxLength 100
   */
  name: string
  /**
   * @maxLength 500
   * @nullable
   */
  description?: string | null
  displayOrder?: number
  isActive?: boolean
  createdAt?: string
  updatedAt?: string
  totalIconGroups?: number
  totalIcons?: number
}